refactor(server): split main.go into focused files by concern
